+++ src/codec/unsuback.rs
@@ -175,7 +175,7 @@ mod test {
 
     #[test]
     fn from_bytes_0() {
-        const FIXED_HDR: u8 = ((UnsubackRx::PACKET_ID as u8) << 4) as u8;
+        const FIXED_HDR: u8 = UnsubackRx::PACKET_ID << 4;
         const PACKET: [u8; 24] = [
             FIXED_HDR,
             22,   // Remaining length
